+++ classes/rental.rb
