+++ classes/student.rb
