+++ classes/teacher.rb
