+++ creators/teacher.rb
@@ -1,0 +1,8 @@
+require_relative '../classes/teacher'
+
+def create_teacher(specialization, person)
+  teacher = Teacher.new(specialization)
+  person[:role] = 'teacher'
+  person[:specialization] = teacher.specialization
+  person
+end
