--- runner.rb
@@ -1,7 +1,0 @@
-require './app'
-require './text'
-
-class Menu
-  app = App.new
-  app.start
-end
